Apply requested changes and add cache to Switch
