+++ src/platform/file_manager.c
@@ -4,6 +4,7 @@
 #include "core/log.h"
 #include "core/string.h"
 #include "platform/android/android.h"
+#include "platform/file_manager_cache.h"
 #include "platform/vita/vita.h"
 
 #include <dirent.h>
@@ -20,7 +21,6 @@
 #define fs_dir_close _wclosedir
 #define fs_dir_read _wreaddir
 #define dir_entry_name(d) wchar_to_utf8(d->d_name)
-#define dir_to_utf8(d) wchar_to_utf8(d)
 typedef const wchar_t * dir_name;
 
 static const char *wchar_to_utf8(const wchar_t *str)
@@ -52,7 +52,6 @@ static wchar_t *utf8_to_wchar(const char *str)
 #define fs_dir_close closedir
 #define fs_dir_read readdir
 #define dir_entry_name(d) ((d)->d_name)
-#define dir_to_utf8(d) (d)
 typedef const char * dir_name;
 #endif
 
@@ -85,154 +84,7 @@ typedef const char * dir_name;
 #include <unistd.h>
 #endif
 
-#ifdef __vita__
-typedef struct file_info {
-    char name[FILE_NAME_MAX];
-    const char *extension;
-    int type;
-    struct file_info *next;
-} file_info;
-
-typedef struct dir_info {
-    char name[FILE_NAME_MAX];
-    file_info *first_file;
-    struct dir_info *next;
-} dir_info;
-
-static dir_info *base_dir_info;
-
-static const dir_info *get_or_create_dir_info(dir_name dir)
-{
-    dir_info *info = base_dir_info;
-    const char *dir_utf8 = dir_to_utf8(dir);
-    while (info) {
-        if (strcmp(info->name, dir_utf8) == 0) {
-            return info;
-        }
-        if (!info->next) {
-            break;
-        }
-        info = info->next;
-    }
-    fs_dir_type *d = fs_dir_open(dir);
-    if (!d) {
-        return 0;
-    }
-    if (!info) {
-        info = malloc(sizeof(dir_info));
-        base_dir_info = info;
-    } else {
-        info->next = malloc(sizeof(dir_info));
-        info = info->next;
-    }
-    strncpy(info->name, dir_utf8, FILE_NAME_MAX - 1);
-    info->name[FILE_NAME_MAX - 1] = 0;
-    info->first_file = 0;
-    info->next = 0;
-    fs_dir_entry *entry;
-    file_info *file_item = 0;
-    int dir_name_offset = 0;
-    while ((entry = fs_dir_read(d))) {
-        const char *name = dir_entry_name(entry);
-        if (name[0] == '.') {
-            // Skip hidden files
-            continue;
-        }
-        if (!file_item) {
-            file_item = malloc(sizeof(file_info));
-            info->first_file = file_item;
-        } else {
-            file_item->next = malloc(sizeof(file_info));
-            file_item = file_item->next;
-        }
-        file_item->next = 0;
-
-        // Copy name
-        strncpy(file_item->name, name, FILE_NAME_MAX - 1);
-        file_item->name[FILE_NAME_MAX - 1] = 0;
-
-        // Copy extension
-        char c;
-        const char *extension = file_item->name;
-        do {
-            c = *extension;
-            extension++;
-        } while (c != '.' && c);
-        file_item->extension = extension;
-
-        // Check type
-        int type = TYPE_FILE;
-        // Stat does not work for Vita, so we check if a file is a directory by trying to open it as a dir
-        // For performance reasons, we only check for a directory if the name has no extension
-        // This is effectively a hack, and definitely not full-proof, but the performance gains are well worth it
-        if (!c) {
-            static char full_name[FILE_NAME_MAX];
-            if (!dir_name_offset) {
-                strncpy(full_name, info->name, FILE_NAME_MAX);
-                dir_name_offset = strlen(info->name);
-            }
-            strncpy(full_name + dir_name_offset, name, FILE_NAME_MAX - 1 - dir_name_offset);
-            fs_dir_type *file_d = fs_dir_open(full_name);
-            if (file_d) {
-                type = TYPE_DIR;
-                fs_dir_close(file_d);
-            }
-        }
-        file_item->type = type;
-    }
-    fs_dir_close(d);
-    return info;
-}
-
-static void add_file_info_to_cache(const char *filename, const char *mode)
-{
-    // Julius only creates files to the base dir
-    if (base_dir_info && strchr(mode, 'w') && !file_exists(filename, NOT_LOCALIZED)) {
-        file_info *f = malloc(sizeof(file_info));
-        strncpy(f->name, filename, FILE_NAME_MAX - 1);
-        f->name[FILE_NAME_MAX - 1] = 0;
-        f->type = TYPE_FILE;
-        char c;
-        const char *name = f->name;
-        do {
-            c = *name;
-            name++;
-        } while (c != '.' && c);
-        f->extension = name;
-        f->next = base_dir_info->first_file;
-        base_dir_info->first_file = f;
-    }
-}
-
-static void delete_file_info_from_cache(const char *filename)
-{
-    // Julius only deletes files from the base dir
-    if (!base_dir_info) {
-        return;
-    }
-    file_info *prev = 0;
-    for (file_info *f = base_dir_info->first_file; f; f = f->next) {
-        if (strcmp(filename, f->name) == 0) {
-            if (prev) {
-                prev->next = f->next;
-            } else {
-                base_dir_info->first_file = f->next;
-            }
-            free(f);
-            return;
-        }
-        prev = f;
-    }
-}
-
-static int file_info_has_extension(const file_info *f, const char *extension)
-{
-    if (!(f->type & TYPE_FILE) || !extension || !*extension) {
-        return 1;
-    }
-    return string_compare_case_insensitive(f->extension, extension) == 0;
-}
-#else
+#if !defined(__vita__) && !defined(__SWITCH__)
 static int is_file(int mode)
 {
     return S_ISREG(mode) || S_ISLNK(mode);
@@ -255,8 +107,8 @@ int platform_file_manager_list_directory_contents(
     }
 #ifdef __ANDROID__
     int match = android_get_directory_contents(current_dir, type, extension, callback);
-#elif defined(__vita__)
-    const dir_info *d = get_or_create_dir_info(current_dir);
+#elif defined(__vita__) || defined(__SWITCH__)
+    const dir_info *d = platform_file_manager_cache_get_dir_info(current_dir);
     if (!d) {
         return LIST_ERROR;
     }
@@ -265,7 +117,7 @@ int platform_file_manager_list_directory_contents(
         if (!(type & f->type)) {
             continue;
         }
-        if (!file_info_has_extension(f, extension)) {
+        if (!platform_file_manager_cache_file_has_extension(f, extension)) {
             continue;
         }
         match = callback(f->name);
@@ -335,11 +187,12 @@ int platform_file_manager_set_base_path(const char *path)
 #endif
 }
 
-#if defined(__vita__)
-
+#ifdef __vita__
 FILE *platform_file_manager_open_file(const char *filename, const char *mode)
 {
-    add_file_info_to_cache(filename, mode);
+    if (strchr(mode, 'w') && !file_exists(filename, NOT_LOCALIZED)) {
+        platform_file_manager_cache_add_file_info(filename);
+    }
     char *resolved_path = vita_prepend_path(filename);
     FILE *fp = fopen(resolved_path, mode);
     free(resolved_path);
@@ -348,7 +201,7 @@ FILE *platform_file_manager_open_file(const char *filename, const char *mode)
 
 int platform_file_manager_remove_file(const char *filename)
 {
-    delete_file_info_from_cache(filename);
+    platform_file_manager_cache_delete_file_info(filename);
     char *resolved_path = vita_prepend_path(filename);
     int result = remove(resolved_path);
     free(resolved_path);
@@ -398,11 +251,19 @@ int platform_file_manager_remove_file(const char *filename)
 
 FILE *platform_file_manager_open_file(const char *filename, const char *mode)
 {
+#ifdef __SWITCH__
+    if (strchr(mode, 'w') && !file_exists(filename, NOT_LOCALIZED)) {
+        platform_file_manager_cache_add_file_info(filename);
+    }
+#endif
     return fopen(filename, mode);
 }
 
 int platform_file_manager_remove_file(const char *filename)
 {
+#ifdef __SWITCH__
+    platform_file_manager_cache_delete_file_info(filename);
+#endif
     return remove(filename) == 0;
 }
